Do a bunch of exercises

+++ 06-variables/05-type-conversion/exercises/05-convert-and-fix-5/main.go
@@ -29,5 +29,7 @@ func main() {
 	max := int16(1000)
 
 	// FIX THE CODE HERE
-	fmt.Println(int8(max) + min)
+	// this also works
+	//fmt.Println(int(max) + int(min))
+	fmt.Println(max + int16(min))
 }
